create message completed

+++ src/main/java/com/example/repository/AccountRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.example.entity.*;
 
 @Repository
-public interface AccountRepository extends JpaRepository<Account, Long> {
+public interface AccountRepository extends JpaRepository<Account, Integer> {
     Optional<Account> getAccountByUsername(String username);
     Optional<Account> findByUsernameAndPassword(String username, String password);
 }
